refactor: extract checkpoint helper
